:recycle: (870): fold s3_service into infrastructure/s3 + add inter-task artifact helpers
